# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SP MLM Platform API",
        "description": "API Documentation for Mobile App and Frontend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Main Server"
        }
    ],
    "paths": {
        "/api/user/account/bind": {
            "get": {
                "tags": [
                    "User Account"
                ],
                "summary": "Get saved Bank and Crypto details",
                "operationId": "74e1ccd5a2309c3032541db6f9ae3a90",
                "responses": {
                    "200": {
                        "description": "Account details fetched successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Account"
                ],
                "summary": "Bind or Update Bank/Crypto Account",
                "operationId": "124f249eeb81a27c975cfb1b2b13a4c8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "bank",
                                            "crypto"
                                        ],
                                        "example": "bank"
                                    },
                                    "bank_name": {
                                        "type": "string",
                                        "example": "State Bank of India (Only for Bank)"
                                    },
                                    "account_holder": {
                                        "type": "string",
                                        "example": "Bipin Sahani (Only for Bank)"
                                    },
                                    "account_number": {
                                        "type": "string",
                                        "example": "1234567890 (Only for Bank)"
                                    },
                                    "ifsc_code": {
                                        "type": "string",
                                        "example": "SBIN0001234 (Only for Bank)"
                                    },
                                    "network": {
                                        "type": "string",
                                        "example": "TRC20 (Only for Crypto)"
                                    },
                                    "wallet_address": {
                                        "type": "string",
                                        "example": "Txyz123abc... (Only for Crypto)"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Account updated successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/register": {
            "post": {
                "tags": [
                    "User Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "be89b85b0616556577ebccb0fe9f5ad3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "mobile",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Bipin Sahani"
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "rid": {
                                        "type": "string",
                                        "example": "SP123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "tags": [
                    "User Authentication"
                ],
                "summary": "Login to get access token",
                "operationId": "4af7f504cb781537c1d3a823bc957e31",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login_input",
                                    "password"
                                ],
                                "properties": {
                                    "login_input": {
                                        "type": "string",
                                        "example": "9876543210"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "403": {
                        "description": "Account is blocked"
                    }
                }
            }
        },
        "/api/user/profile": {
            "get": {
                "tags": [
                    "User Authentication"
                ],
                "summary": "Get logged in user profile details",
                "operationId": "54f7586f0f4778851d73823446817a0a",
                "responses": {
                    "200": {
                        "description": "Profile fetched successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/logout": {
            "post": {
                "tags": [
                    "User Authentication"
                ],
                "summary": "Logout and invalidate token",
                "operationId": "11166c57e2349b46882caa0a44ca5778",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/dashboard": {
            "get": {
                "tags": [
                    "User Dashboard"
                ],
                "summary": "Get User Dashboard Data",
                "operationId": "a3d0e2e37804a1f69175261544162d56",
                "responses": {
                    "200": {
                        "description": "Dashboard data fetched successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/level/{id}/join": {
            "post": {
                "tags": [
                    "User Dashboard"
                ],
                "summary": "Join or Upgrade a Level",
                "operationId": "d60ab986953d18bba7a3e26a8ff041b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the level to join",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully joined/upgraded"
                    },
                    "400": {
                        "description": "Business logic error (Insufficient balance, etc.)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user/guide": {
            "get": {
                "tags": [
                    "User Dashboard"
                ],
                "summary": "Get Banners and Income Methods",
                "operationId": "3b868a4219fb6bb9d12eb8e04c72beee",
                "responses": {
                    "200": {
                        "description": "Guide data fetched successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Login with Mobile and Password to get the authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}